Correction of damage variable

+++ Assets/Scripts/BaseProjectile.cs
@@ -6,6 +6,6 @@ public abstract class BaseProjectile : MonoBehaviour
 {
     public float speed = 5.0f;
 
-    public abstract void FireProjectile(GameObject launcher, GameObject target, int damage);
+    public abstract void FireProjectile(GameObject launcher, GameObject target, float damage);
 
 }
